# cmd/server/main.go
package main

import (
	"context"
	"database/sql"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/google/uuid"
	"github.com/liamcoop/rules/multitenantengine"
	"github.com/liamcoop/rules/rules"
	_ "github.com/lib/pq"
	httpSwagger "github.com/swaggo/http-swagger"

	_ "github.com/liamcoop/rules/cmd/server/docs" // Swagger docs
)

// @title Multi-Tenant Rules Engine API
// @version 1.0
// @description A multi-tenant rules engine with dynamic schemas and CEL rule evaluation
// @description
// @description ## Features
// @description - Multi-tenant support with isolated data
// @description - Dynamic schema definition per tenant
// @description - CEL (Common Expression Language) rule engine
// @description - Real-time rule evaluation
// @description - Zero-downtime schema updates
// @description
// @description ## Supported Data Types
// @description - int, int64, float64, string, bool, bytes, timestamp, duration
// @description
// @description ## Authentication
// @description Not yet implemented (planned for production)
//
// @contact.name API Support
// @contact.email support@example.com
//
// @license.name MIT
// @license.url https://opensource.org/licenses/MIT
//
// @host localhost:8080
// @BasePath /
// @schemes http

type Server struct {
	db            *sql.DB
	engineManager *multitenantengine.MultiTenantEngineManager
	router        *chi.Mux
}

func NewServer(databaseURL string) (*Server, error) {
	// Connect to database
	db, err := sql.Open("postgres", databaseURL)
	if err != nil {
		return nil, fmt.Errorf("failed to open database: %w", err)
	}

	if err := db.Ping(); err != nil {
		return nil, fmt.Errorf("failed to ping database: %w", err)
	}

	return NewServerWithDB(db)
}

func NewServerWithDB(db *sql.DB) (*Server, error) {
	// Create engine manager
	engineManager := multitenantengine.NewMultiTenantEngineManager(db)

	// Load all tenants
	log.Println("Loading tenants from database...")
	if err := engineManager.LoadAllTenants(); err != nil {
		return nil, fmt.Errorf("failed to load tenants: %w", err)
	}

	tenants := engineManager.ListTenants()
	log.Printf("Loaded %d tenants: %v", len(tenants), tenants)

	s := &Server{
		db:            db,
		engineManager: engineManager,
	}

	s.setupRoutes()

	return s, nil
}

func (s *Server) setupRoutes() {
	r := chi.NewRouter()

	// Middleware
	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)
	r.Use(middleware.Timeout(60 * time.Second))

	// Swagger documentation
	r.Get("/swagger/*", httpSwagger.Handler(
		httpSwagger.URL("http://localhost:8080/swagger/doc.json"),
	))

	// Health check
	r.Get("/api/v1/health", s.handleHealth)

	// Evaluation
	r.Post("/api/v1/evaluate", s.handleEvaluate)

	// Tenant management
	r.Route("/api/v1/tenants", func(r chi.Router) {
		r.Get("/", s.handleListTenants)
		r.Post("/", s.handleCreateTenant)

		r.Route("/{tenantId}", func(r chi.Router) {
			// Schema management
			r.Post("/schema", s.handleCreateSchema)
			r.Put("/schema", s.handleUpdateSchema)
			r.Get("/schema", s.handleGetSchema)

			// Rule management
			r.Post("/rules", s.handleCreateRule)
			r.Get("/rules", s.handleListRules)
			r.Get("/rules/{ruleId}", s.handleGetRule)
			r.Put("/rules/{ruleId}", s.handleUpdateRule)
			r.Delete("/rules/{ruleId}", s.handleDeleteRule)
		})
	})

	s.router = r
}

func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	s.router.ServeHTTP(w, r)
}

// handleHealth godoc
// @Summary Health check
// @Description Check if the service and database are healthy
// @Tags health
// @Produce json
// @Success 200 {object} HealthResponse
// @Failure 503 {object} ErrorResponse
// @Router /api/v1/health [get]
func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	if err := s.db.Ping(); err != nil {
		respondJSON(w, http.StatusServiceUnavailable, map[string]string{
			"status": "unhealthy",
			"error":  err.Error(),
		})
		return
	}

	respondJSON(w, http.StatusOK, map[string]any{
		"status":        "healthy",
		"tenantsLoaded": len(s.engineManager.ListTenants()),
	})
}

// handleEvaluate godoc
// @Summary Evaluate rules against facts
// @Description Evaluate a tenant's rules against provided facts. Returns which rules matched.
// @Tags evaluation
// @Accept json
// @Produce json
// @Param request body EvaluateRequest true "Evaluation request with tenant ID, facts, and optional rule IDs"
// @Success 200 {object} EvaluateResponse
// @Failure 400 {object} ErrorResponse "Invalid request or facts don't match schema"
// @Failure 404 {object} ErrorResponse "Tenant not found"
// @Failure 500 {object} ErrorResponse "Evaluation error"
// @Router /api/v1/evaluate [post]
func (s *Server) handleEvaluate(w http.ResponseWriter, r *http.Request) {
	var req struct {
		TenantID string         `json:"tenantId"`
		Facts    map[string]any `json:"facts"`
		RuleIDs  []string       `json:"rules,omitempty"` // optional
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body", err)
		return
	}

	if req.TenantID == "" {
		respondError(w, http.StatusBadRequest, "tenantId is required", nil)
		return
	}

	if req.Facts == nil {
		respondError(w, http.StatusBadRequest, "facts are required", nil)
		return
	}

	// Get tenant's engine
	engine, err := s.engineManager.GetEngine(req.TenantID)
	if err != nil {
		respondError(w, http.StatusNotFound, "tenant not found", err)
		return
	}

	startTime := time.Now()

	// Evaluate rules
	var results []*rules.EvaluationResult
	if len(req.RuleIDs) > 0 {
		// Evaluate specific rules
		results = make([]*rules.EvaluationResult, 0, len(req.RuleIDs))
		for _, ruleID := range req.RuleIDs {
			result, err := engine.Evaluate(ruleID, req.Facts)
			if err != nil {
				// Continue on error (might be rule not found)
				log.Printf("Error evaluating rule %s: %v", ruleID, err)
				continue
			}
			results = append(results, result)
		}
	} else {
		// Evaluate all active rules
		results, err = engine.EvaluateAll(req.Facts)
		if err != nil {
			respondError(w, http.StatusInternalServerError, "evaluation failed", err)
			return
		}
	}

	evaluationTime := time.Since(startTime)

	// Format response
	response := map[string]any{
		"results":        results,
		"evaluationTime": evaluationTime.String(),
	}

	respondJSON(w, http.StatusOK, response)
}

// List tenants handler
func (s *Server) handleListTenants(w http.ResponseWriter, r *http.Request) {
	rows, err := s.db.Query("SELECT id, name, created_at, updated_at FROM tenants ORDER BY created_at DESC")
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to list tenants", err)
		return
	}
	defer rows.Close()

	type tenant struct {
		ID        string    `json:"id"`
		Name      string    `json:"name"`
		CreatedAt time.Time `json:"createdAt"`
		UpdatedAt time.Time `json:"updatedAt"`
	}

	tenants := []tenant{}
	for rows.Next() {
		var t tenant
		if err := rows.Scan(&t.ID, &t.Name, &t.CreatedAt, &t.UpdatedAt); err != nil {
			respondError(w, http.StatusInternalServerError, "failed to scan tenant", err)
			return
		}
		tenants = append(tenants, t)
	}

	respondJSON(w, http.StatusOK, map[string]any{
		"tenants": tenants,
	})
}

// handleCreateTenant godoc
// @Summary Create a new tenant
// @Description Create a new tenant in the system
// @Tags tenants
// @Accept json
// @Produce json
// @Param tenant body CreateTenantRequest true "Tenant details"
// @Success 201 {object} TenantResponse
// @Failure 400 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /api/v1/tenants [post]
func (s *Server) handleCreateTenant(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Name string `json:"name"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body", err)
		return
	}

	if req.Name == "" {
		respondError(w, http.StatusBadRequest, "name is required", nil)
		return
	}

	var tenantID string
	err := s.db.QueryRow(`
		INSERT INTO tenants (name, created_at, updated_at)
		VALUES ($1, NOW(), NOW())
		RETURNING id
	`, req.Name).Scan(&tenantID)

	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to create tenant", err)
		return
	}

	respondJSON(w, http.StatusCreated, map[string]any{
		"id":   tenantID,
		"name": req.Name,
	})
}

// handleCreateSchema godoc
// @Summary Create a schema for a tenant
// @Description Create a new schema definition for a tenant. Can only be called once per tenant. See validation rules in documentation.
// @Tags schemas
// @Accept json
// @Produce json
// @Param tenantId path string true "Tenant ID"
// @Param schema body CreateSchemaRequest true "Schema definition"
// @Success 201 {object} SchemaResponse
// @Failure 400 {object} ErrorResponse "Invalid schema - see validation rules"
// @Failure 404 {object} ErrorResponse "Tenant not found"
// @Failure 409 {object} ErrorResponse "Schema already exists"
// @Router /api/v1/tenants/{tenantId}/schema [post]
func (s *Server) handleCreateSchema(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")

	var req struct {
		Definition multitenantengine.Schema `json:"definition"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body", err)
		return
	}

	if req.Definition == nil {
		respondError(w, http.StatusBadRequest, "definition is required", nil)
		return
	}

	// REQ-SEC-003: Validate schema before any database operations
	// REQ-API-001: Create endpoint SHALL validate schema
	if err := multitenantengine.ValidateSchema(req.Definition); err != nil {
		respondError(w, http.StatusBadRequest, fmt.Sprintf("schema validation failed: %v", err), nil)
		return
	}

	// Check if tenant exists
	var exists bool
	err := s.db.QueryRow("SELECT EXISTS(SELECT 1 FROM tenants WHERE id = $1)", tenantID).Scan(&exists)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to check tenant", err)
		return
	}
	if !exists {
		respondError(w, http.StatusNotFound, "tenant not found", nil)
		return
	}

	// Check if schema already exists
	var schemaExists bool
	err = s.db.QueryRow("SELECT EXISTS(SELECT 1 FROM schemas WHERE tenant_id = $1)", tenantID).Scan(&schemaExists)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to check schema", err)
		return
	}
	if schemaExists {
		respondError(w, http.StatusConflict, "schema already exists, use PUT to update", nil)
		return
	}

	// Create schema in database
	schemaJSON, err := json.Marshal(req.Definition)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to marshal schema", err)
		return
	}

	var version int
	err = s.db.QueryRow(`
		INSERT INTO schemas (tenant_id, version, definition, active, created_at)
		VALUES ($1, 1, $2, true, NOW())
		RETURNING version
	`, tenantID, schemaJSON).Scan(&version)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to create schema", err)
		return
	}

	// Load tenant engine
	err = s.engineManager.CreateTenant(tenantID, req.Definition)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to load tenant engine", err)
		return
	}

	respondJSON(w, http.StatusCreated, map[string]any{
		"version":    version,
		"status":     "active",
		"definition": req.Definition,
	})
}

// Update schema handler
func (s *Server) handleUpdateSchema(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")

	var req struct {
		Definition multitenantengine.Schema `json:"definition"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body", err)
		return
	}

	// REQ-SEC-003: Validate schema before any database operations
	// REQ-API-002: Update endpoint SHALL validate schema
	if err := multitenantengine.ValidateSchema(req.Definition); err != nil {
		respondError(w, http.StatusBadRequest, fmt.Sprintf("schema validation failed: %v", err), nil)
		return
	}

	// Update schema (zero downtime!)
	err := s.engineManager.UpdateTenantSchema(tenantID, req.Definition)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to update schema", err)
		return
	}

	// Get the new schema version
	var version int
	var schemaJSON []byte
	err = s.db.QueryRow(`
		SELECT version, definition FROM schemas
		WHERE tenant_id = $1 AND active = true
	`, tenantID).Scan(&version, &schemaJSON)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to get schema version", err)
		return
	}

	respondJSON(w, http.StatusOK, map[string]any{
		"version":    version,
		"status":     "active",
		"definition": req.Definition,
	})
}

// Get schema handler
func (s *Server) handleGetSchema(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")

	var schemaJSON []byte
	var version int
	err := s.db.QueryRow(`
		SELECT version, definition
		FROM schemas
		WHERE tenant_id = $1 AND active = true
	`, tenantID).Scan(&version, &schemaJSON)

	if err == sql.ErrNoRows {
		respondError(w, http.StatusNotFound, "schema not found", nil)
		return
	}
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to get schema", err)
		return
	}

	var schema multitenantengine.Schema
	if err := json.Unmarshal(schemaJSON, &schema); err != nil {
		respondError(w, http.StatusInternalServerError, "failed to parse schema", err)
		return
	}

	respondJSON(w, http.StatusOK, map[string]any{
		"version":    version,
		"definition": schema,
	})
}

// Create rule handler
func (s *Server) handleCreateRule(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")

	var req struct {
		Name       string `json:"name"`
		Expression string `json:"expression"`
		Active     bool   `json:"active"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body", err)
		return
	}

	if req.Name == "" || req.Expression == "" {
		respondError(w, http.StatusBadRequest, "name and expression are required", nil)
		return
	}

	// Get engine to add and compile the rule
	engine, err := s.engineManager.GetEngine(tenantID)
	if err != nil {
		respondError(w, http.StatusNotFound, "tenant not found", err)
		return
	}

	// Create rule
	rule := &rules.Rule{
		Name:       req.Name,
		Expression: req.Expression,
		Active:     req.Active,
		CreatedAt:  time.Now(),
		UpdatedAt:  time.Now(),
	}

	// Generate UUID for rule ID
	rule.ID = generateUUID()

	// Add rule (this validates and compiles it)
	if err := engine.AddRule(rule); err != nil {
		respondError(w, http.StatusBadRequest, "failed to add rule", err)
		return
	}

	respondJSON(w, http.StatusCreated, map[string]any{
		"id":         rule.ID,
		"name":       rule.Name,
		"expression": rule.Expression,
		"active":     rule.Active,
	})
}

// List rules handler
func (s *Server) handleListRules(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")

	rows, err := s.db.Query(`
		SELECT id, name, expression, active, created_at, updated_at
		FROM rules
		WHERE tenant_id = $1
		ORDER BY created_at DESC
	`, tenantID)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to list rules", err)
		return
	}
	defer rows.Close()

	rulesList := []*rules.Rule{}
	for rows.Next() {
		var r rules.Rule
		if err := rows.Scan(&r.ID, &r.Name, &r.Expression, &r.Active, &r.CreatedAt, &r.UpdatedAt); err != nil {
			respondError(w, http.StatusInternalServerError, "failed to scan rule", err)
			return
		}
		rulesList = append(rulesList, &r)
	}

	respondJSON(w, http.StatusOK, map[string]any{
		"rules": rulesList,
	})
}

// Get rule handler
func (s *Server) handleGetRule(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")
	ruleID := chi.URLParam(r, "ruleId")

	store := rules.NewPostgresRuleStore(s.db, tenantID)
	rule, err := store.Get(ruleID)
	if err != nil {
		respondError(w, http.StatusNotFound, "rule not found", err)
		return
	}

	respondJSON(w, http.StatusOK, rule)
}

// Update rule handler
func (s *Server) handleUpdateRule(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")
	ruleID := chi.URLParam(r, "ruleId")

	var req struct {
		Name       string `json:"name"`
		Expression string `json:"expression"`
		Active     bool   `json:"active"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "invalid request body", err)
		return
	}

	// Get engine
	engine, err := s.engineManager.GetEngine(tenantID)
	if err != nil {
		respondError(w, http.StatusNotFound, "tenant not found", err)
		return
	}

	// Update rule
	rule := &rules.Rule{
		ID:         ruleID,
		Name:       req.Name,
		Expression: req.Expression,
		Active:     req.Active,
		UpdatedAt:  time.Now(),
	}

	if err := engine.UpdateRule(rule); err != nil {
		respondError(w, http.StatusBadRequest, "failed to update rule", err)
		return
	}

	respondJSON(w, http.StatusOK, rule)
}

// Delete rule handler
func (s *Server) handleDeleteRule(w http.ResponseWriter, r *http.Request) {
	tenantID := chi.URLParam(r, "tenantId")
	ruleID := chi.URLParam(r, "ruleId")

	engine, err := s.engineManager.GetEngine(tenantID)
	if err != nil {
		respondError(w, http.StatusNotFound, "tenant not found", err)
		return
	}

	if err := engine.DeleteRule(ruleID); err != nil {
		respondError(w, http.StatusNotFound, "rule not found", err)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// Helper functions
func respondJSON(w http.ResponseWriter, status int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(data)
}

func respondError(w http.ResponseWriter, status int, message string, err error) {
	response := map[string]string{
		"error": message,
	}
	if err != nil {
		response["details"] = err.Error()
	}
	respondJSON(w, status, response)
}

func generateUUID() string {
	return uuid.New().String()
}

func main() {
	// Get database URL from environment
	databaseURL := os.Getenv("DATABASE_URL")
	if databaseURL == "" {
		log.Fatal("DATABASE_URL environment variable is required")
	}

	// Create server
	server, err := NewServer(databaseURL)
	if err != nil {
		log.Fatalf("Failed to create server: %v", err)
	}
	defer server.db.Close()

	// Start HTTP server
	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	httpServer := &http.Server{
		Addr:         ":" + port,
		Handler:      server,
		ReadTimeout:  15 * time.Second,
		WriteTimeout: 15 * time.Second,
		IdleTimeout:  60 * time.Second,
	}

	// Graceful shutdown handling
	go func() {
		log.Printf("Server starting on port %s", port)
		if err := httpServer.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Fatalf("Server failed to start: %v", err)
		}
	}()

	// Wait for interrupt signal
	sigChan := make(chan os.Signal, 1)
	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
	<-sigChan

	log.Println("Shutting down server...")
	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
	defer cancel()

	if err := httpServer.Shutdown(ctx); err != nil {
		log.Printf("Server shutdown error: %v", err)
	}

	log.Println("Server stopped")
}